Fix keras docstrings
PiperOrigin-RevId: 424275818

+++ keras/layers/preprocessing/category_encoding.py
@@ -80,7 +80,7 @@ class CategoryEncoding(base_layer.Layer):
     array([[0.1, 0.2, 0. , 0. ],
            [0.2, 0. , 0. , 0. ],
            [0. , 0.2, 0.3, 0. ],
-           [0. , 0.2, 0. , 0.4]])>
+           [0. , 0.2, 0. , 0.4]], dtype=float32)>
 
   Args:
     num_tokens: The total number of tokens the layer should support. All inputs
